+++ changelog.md
@@ -5,11 +5,16 @@
 ### Added
 
 - Added support for Secure Custom Fields detection
+- Added link to support forum on plugin in plugin list
 
 ### Optimized
 
 - Log database errors
 - Renamed support object Secure Custom Fields to Advanced Custom Fields
+
+### Fixed
+
+- Fixed wrong usage of plugin name as ID in backend
 
 ## [4.2.3] - 18.02.2025